requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
linkify-it-py==2.0.2
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
openai==0.28.0
Pygments==2.16.1
python-dotenv==1.0.0
requests==2.31.0
rich==13.5.3
textual==0.37.1
tqdm==4.66.1
typing_extensions==4.7.1
uc-micro-py==1.0.2
urllib3==2.0.4
yarl==1.9.2
zipp==3.16.2
